+++ CHANGES.md
@@ -5,7 +5,7 @@ Release Notes.
 9.7.0
 ------------------
 
-
+* Added support for Lettuce reactive Redis commands.
 
 All issues and pull requests are [here](https://github.com/apache/skywalking/milestone/249?closed=1)